Fixed provisioning profiles shit.

+++ Wigo/MainViewController.m
@@ -385,7 +385,7 @@
 - (UIImageView *)gifGoOut {
     NSMutableArray *goOutArray = [[NSMutableArray array] init];
     for (NSUInteger i  = 1; i <= 60; i++) {
-        NSString *fileName = [NSString stringWithFormat:@"go-out_2-%d.png",i];
+        NSString *fileName = [NSString stringWithFormat:@"go-out_2-%lu.png",(unsigned long)i];
         [goOutArray addObject:[UIImage imageNamed:fileName]];
     }
     UIImageView *gifGoOutImageView = [[UIImageView alloc] init];
@@ -397,7 +397,7 @@
 - (UIImageView *)gifGlowing {
     NSMutableArray *goOutArray = [[NSMutableArray array] init];
     for (NSUInteger i  = 0; i <= 59; i++) {
-        NSString *fileName = [NSString stringWithFormat:@"glowing-%d.png",i];
+        NSString *fileName = [NSString stringWithFormat:@"glowing-%lu.png",(unsigned long)i];
         [goOutArray addObject:[UIImage imageNamed:fileName]];
     }
     UIImageView *gifGoOutImageView = [[UIImageView alloc] init];
